p4i integration Integrate 217450 from 4.1-optimized to main: Integrate 217449 from 4.1 to 4.1-optimized: Fixes:    Fix for a compiler bug on the Intel compiler. Task:     111144 RevBy:    Kent Details:  The previous implementation of speedAdjustedDelay() always returned 0 in release mode due to a compiler bug on the Intel compiler 9,        and I see no reason to use float. (Actually, it is hardly any reason to use qint64 either, the only argument is to reduce the risk of        overflow. Due to that we do our math in qint64 instead of float.)
[git-p4: depot-paths = "//depot/qt/main/": change = 217451]

+++ src/gui/image/qmovie.cpp
@@ -285,7 +285,7 @@ bool QMoviePrivate::isDone()
 */
 int QMoviePrivate::speedAdjustedDelay(int delay) const
 {
-    return int(float(delay) * float(100.0) / float(speed));
+    return int( (qint64(delay) * qint64(100) ) / qint64(speed) );
 }
 
 /*!
